Rename to shelly_mqtt
