+++ CODEOWNERS
@@ -353,7 +353,7 @@ homeassistant/components/serial/* @fabaff
 homeassistant/components/seven_segments/* @fabaff
 homeassistant/components/seventeentrack/* @bachya
 homeassistant/components/shell_command/* @home-assistant/core
-homeassistant/components/shelly/* @OnFreund
+homeassistant/components/shelly_mqtt/* @OnFreund
 homeassistant/components/shiftr/* @fabaff
 homeassistant/components/shodan/* @fabaff
 homeassistant/components/sighthound/* @robmarkcole
