--- homeassistant/components/shelly/manifest.json
@@ -1,9 +1,0 @@
-{
-  "domain": "shelly",
-  "name": "Shelly",
-  "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/shelly",
-  "requirements": [],
-  "after_dependencies": ["mqtt"],
-  "codeowners": ["@OnFreund"]
-}
