--- tests/components/shelly/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Shelly integration."""
