+++ tests/components/shelly_mqtt/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Shelly MQTT integration."""
